Updates to modularize the javascript code
