+++ web_interface/html/index.html
@@ -26,6 +26,6 @@
   <button onclick="">Apply</button>
 
 </div>
-  <script src="main.js"></script>
+  <script type="module"  src="../javascript/main.js"></script>
 </body>
 </html>
